# mfdc-nc/docs/swagger.yaml
definitions:
...
  model.Pool:
    properties:
      active:
        type: boolean
      block:
        type: integer
      created_at:
        type: string
      finish:
        type: boolean
      finish_at:
        type: string
      id:
        type: integer
      name:
        type: string
      num_count:
        type: integer
      rotation:
        type: boolean
    type: object
...
  model.Scheduler:
    properties:
      id:
        type: integer
      name:
        type: string
      periodic_sec:
        type: integer
      start_time:
        $ref: '#/definitions/model.TimeTZ'
      stop_time:
        $ref: '#/definitions/model.TimeTZ'
      team_id:
        type: integer
      team_name:
        type: string
    type: object
...
  model.SwaggerDefaultResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  model.SwaggerPoolsList:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Pool'
        type: array
      status:
        type: string
    type: object
...
  model.TimeTZ:
    properties:
      time:
        type: string
    type: object
...
info:
  contact: {}
  description: Swagger API for Golang Project MFDC NC
  title: MFDC Number changer
  version: "1.0"
paths:
...
  /pools/{id}/activate:
    get:
      consumes:
      - application/json
      description: Activate pool
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Activate pool
      tags:
      - Pools
  /pools/{id}/deactivate:
    get:
      consumes:
      - application/json
      description: Deactivate pool
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Deactivate pool
      tags:
      - Pools
  /pools/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Drop pool
      parameters:
      - description: Pool ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Delete pool
      tags:
      - Pools
  /pools/list:
    get:
      consumes:
      - application/json
      description: List pools
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerPoolsList'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List pools
      tags:
      - Pools
  /schedule/add:
    post:
      consumes:
      - application/json
      description: Add new schedule
      parameters:
      - description: 'Schedule, time format: 00:00:00+03'
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/model.Scheduler'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Add scheduler
      tags:
      - Scheduler
  /schedule/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Drop schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Delete schedule
      tags:
      - Scheduler
  /schedule/edit/{id}:
    patch:
      consumes:
      - application/json
      description: Edit schedule params
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data without ID
        in: body
        name: ip
        required: true
        schema:
          $ref: '#/definitions/model.Scheduler'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Edt schedule
      tags:
      - Scheduler
  /schedule/list:
    get:
      consumes:
      - application/json
      description: List scheduls
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Scheduler'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List scheduls
      tags:
      - Scheduler
  /subpools/{pool_id}/next:
    get:
      consumes:
      - application/json
      description: Manual activate new subpool for pool
      parameters:
      - description: Pool ID
        in: path
        name: pool_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Manual activate next subpool for pool
      tags:
      - Subpools
  /subpools/next:
    get:
      consumes:
      - application/json
      description: Manual activate new subpool
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SwaggerDefaultResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Manual activate next subpool
      tags:
      - Subpools
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-MFDC-Key
    type: apiKey
swagger: "2.0"